feat: load pkl (csml) module from module path
