+++ src/test/java/at/ac/uibk/dps/cirrina/csml/parser/CsmlParserTest.java
@@ -1,0 +1,27 @@
+package at.ac.uibk.dps.cirrina.csml.parser;
+
+import org.junit.jupiter.api.Test;
+
+class CsmlParserTest {
+
+  @Test
+  void testDescriptionPositive() {
+    // TODO: Fix me
+    /*var pkl = DefaultDescriptions.complete;
+
+    assertDoesNotThrow(() -> {
+      var csm = CsmlParser.parse(pkl);
+    });*/
+  }
+
+  @Test
+  void testDescriptionNegative() {
+    // TODO: Fix me
+    /*var json = DefaultDescriptions.empty;
+
+    assertThrows(IllegalArgumentException.class, () -> {
+      var csm = CsmlParser.parse(json);
+      System.out.println(csm);
+    });*/
+  }
+}
