--- src/test/java/at/ac/uibk/dps/cirrina/csml/parser/DescriptionParserTest.java
@@ -1,33 +1,0 @@
-package at.ac.uibk.dps.cirrina.csml.parser;
-
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import at.ac.uibk.dps.cirrina.csml.description.Csml;
-import at.ac.uibk.dps.cirrina.data.DefaultDescriptions;
-import at.ac.uibk.dps.cirrina.io.description.DescriptionParser;
-import org.junit.jupiter.api.Test;
-
-class DescriptionParserTest {
-
-  @Test
-  void testDescriptionPositive() {
-    var pkl = DefaultDescriptions.complete;
-
-    var parser = new DescriptionParser<>(Csml.class);
-    assertDoesNotThrow(() -> {
-      var csm = parser.parse(pkl);
-    });
-  }
-
-  @Test
-  void testDescriptionNegative() {
-    var json = DefaultDescriptions.empty;
-
-    var parser = new DescriptionParser<>(Csml.class);
-    assertThrows(IllegalArgumentException.class, () -> {
-      var csm = parser.parse(json);
-      System.out.println(csm);
-    });
-  }
-}
