+++ src/test/java/at/ac/uibk/dps/cirrina/io/description/CsmlParserTest.java
@@ -1,0 +1,9 @@
+package at.ac.uibk.dps.cirrina.io.description;
+
+import org.junit.jupiter.api.Test;
+
+public class CsmlParserTest {
+
+  @Test
+  void loadProject() {}
+}
