+++ src/test/java/at/ac/uibk/dps/cirrina/io/plantuml/CollaborativeStateMachineClassExporterTest.java
@@ -4,9 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import at.ac.uibk.dps.cirrina.classes.collaborativestatemachine.CollaborativeStateMachineClass;
 import at.ac.uibk.dps.cirrina.classes.collaborativestatemachine.CollaborativeStateMachineClassBuilder;
-import at.ac.uibk.dps.cirrina.csml.description.Csml;
 import at.ac.uibk.dps.cirrina.data.DefaultDescriptions;
-import at.ac.uibk.dps.cirrina.io.description.DescriptionParser;
+import at.ac.uibk.dps.cirrina.io.description.CsmlParser;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.StringWriter;
@@ -21,8 +20,7 @@ class CollaborativeStateMachineClassExporterTest {
   static void setUp() {
     var json = DefaultDescriptions.completeNested;
 
-    var parser = new DescriptionParser<>(Csml.class);
-    var csm = assertDoesNotThrow(() -> parser.parse(json));
+    var csm = assertDoesNotThrow(() -> CsmlParser.parse(json));
     completeNestedCsm = assertDoesNotThrow(() ->
       CollaborativeStateMachineClassBuilder.from(csm).build()
     );
