+++ src/test/java/at/ac/uibk/dps/cirrina/runtime/offline/TimeoutTest.java
@@ -6,13 +6,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import at.ac.uibk.dps.cirrina.classes.collaborativestatemachine.CollaborativeStateMachineClass;
 import at.ac.uibk.dps.cirrina.classes.collaborativestatemachine.CollaborativeStateMachineClassBuilder;
-import at.ac.uibk.dps.cirrina.csml.description.Csml;
 import at.ac.uibk.dps.cirrina.data.DefaultDescriptions;
 import at.ac.uibk.dps.cirrina.execution.object.context.InMemoryContext;
 import at.ac.uibk.dps.cirrina.execution.object.event.Event;
 import at.ac.uibk.dps.cirrina.execution.object.event.EventHandler;
 import at.ac.uibk.dps.cirrina.execution.service.OptimalServiceImplementationSelector;
-import at.ac.uibk.dps.cirrina.io.description.DescriptionParser;
+import at.ac.uibk.dps.cirrina.io.description.CsmlParser;
 import at.ac.uibk.dps.cirrina.runtime.OfflineRuntime;
 import com.google.common.collect.ArrayListMultimap;
 import java.io.IOException;
@@ -28,10 +27,9 @@ public class TimeoutTest {
   public static void setUp() {
     final var json = DefaultDescriptions.timeout;
 
-    final var parser = new DescriptionParser<>(Csml.class);
     Assertions.assertDoesNotThrow(() -> {
       collaborativeStateMachineClass = CollaborativeStateMachineClassBuilder.from(
-        parser.parse(json)
+        CsmlParser.parse(json)
       ).build();
     });
   }
