--- src/test/resources/pkl/pingPong.pkl
@@ -1,131 +1,0 @@
-amends "modulepath:/pkl/csm/Csml.pkl"
-import "modulepath:/pkl/csm/Csml.pkl"
-
-name = "ping-pong"
-version = "3.0"
-stateMachines {
-  new AStateMachine {}
-  new BStateMachine {}
-}
-
-local class AStateMachine extends Csml.StateMachine {
-  name = "stateMachine1"
-  states {
-    new {
-      name = "a"
-      initial = true
-      entry {
-        new Csml.AssignAction {
-          variable {
-            name = "v"
-            value = "v + 1"
-          }
-        }
-        new Csml.RaiseAction {
-          event {
-            name = "e1"
-            channel = "global"
-          }
-        }
-      }
-      on {
-        new Csml.OnTransition {
-          event = "e2"
-          target = "b"
-        }
-      }
-    }
-    new {
-        name = "b"
-        entry {
-          new Csml.RaiseAction {
-            event {
-              name = "e3"
-              channel = "global"
-            }
-          }
-        }
-        on {
-          new Csml.OnTransition {
-            event = "e4"
-            target = "a"
-          }
-        }
-        always {
-          new Csml.Transition {
-            target = "c"
-            guards {
-              new Csml.Guard {
-                expression = "v >= 100"
-              }
-            }
-          }
-        }
-    }
-    new {
-      name = "c"
-      terminal = true
-    }
-  }
-}
-
-local class BStateMachine extends Csml.StateMachine {
-  name = "stateMachine2"
-  states {
-    new {
-      name = "a"
-      initial = true
-      entry {
-        new Csml.RaiseActionDescription {
-          event {
-            name = "e4"
-            channel = "global"
-          }
-        }
-      }
-      on {
-        new Csml.OnTransitionDescription {
-          event = "e1"
-          target = "b"
-        }
-      }
-    }
-    new {
-      name = "b"
-      entry {
-        new Csml.AssignActionDescription {
-          variable {
-            name = "v"
-            value = "v + 1"
-          }
-        }
-        new Csml.RaiseActionDescription {
-          event {
-            name = "e2"
-            channel = "global"
-          }
-        }
-      }
-      on {
-        new Csml.OnTransitionDescription {
-          event = "e3"
-          target = "a"
-        }
-      }
-      always {
-        new Csml.TransitionDescription {
-          target = "c"
-          guards {
-            new Csml.GuardDescription {
-              expression = "v >= 100"
-            }
-          }
-        }
-      }
-    }
-    new {
-      name = "c"
-      terminal = true
-    }
-  }
-}
