+++ src/test/resources/pkl/pingPong/stateMachineOne.pkl
@@ -1,0 +1,74 @@
+import "modulepath:/pkl/csm/Csml.pkl"
+
+//             State Machine One
+// ┌──────────────────────────────────────────┐
+// │                                          │
+// │                 e2                       │
+// │           ┌────────────┐                 │
+// │          ┌┴┐          ┌▼┐  v >= 100  ┌─┐ │
+// │ raise e1 │A│ raise e3 │B├────────────►C│ │
+// │  v += 1  └▲┘          └┬┘            └─┘ │
+// │           └────────────┘                 │
+// │                 e4                       │
+// │                                          │
+// └──────────────────────────────────────────┘
+class StateMachine extends Csml.StateMachine {
+  name = "stateMachine1"
+  states {
+    new {
+      name = "a"
+      initial = true
+      entry {
+        new Csml.AssignAction {
+          variable {
+            name = "v"
+            value = "v + 1"
+          }
+        }
+        new Csml.RaiseAction {
+          event {
+            name = "e1"
+            channel = "global"
+          }
+        }
+      }
+      on {
+        new Csml.OnTransition {
+          event = "e2"
+          target = "b"
+        }
+      }
+    }
+    new {
+      name = "b"
+      entry {
+        new Csml.RaiseAction {
+          event {
+            name = "e3"
+            channel = "global"
+          }
+        }
+      }
+      on {
+        new Csml.OnTransition {
+          event = "e4"
+          target = "a"
+        }
+      }
+      always {
+        new Csml.Transition {
+          target = "c"
+          guards {
+            new Csml.Guard {
+              expression = "v >= 100"
+            }
+          }
+        }
+      }
+    }
+    new {
+      name = "c"
+      terminal = true
+    }
+  }
+}
